// src/app/shared/Models/catalogos/caseta.model.ts
export class Caseta {
    id: number = 0;
    nombre: string = "";
    aliasIave: string = "";
    descripcion: string = "";
    metodoPagoId : number = 0;
    geocercaId: number = 0;
    nombreGeocerca: string = "";
    bActivo: boolean = true;
    bCasetaActiva: boolean = true;
    nombreMetodoPago: string = "";
    url: string = "";
    archivoTicketFrente: string = "";
    archivoTicketAtras: string = "";

    constructor(){}
}